Google Authentication, Allowing only BC students to Submit
I created a firebase project in order to google authenticate anyone trying to submit. However, they must be logged in with a @bc.edu email. I also was able to patch some initial errors. Now users only have to authenticate one time successfully and it will remember them. It is still hard to test other emails, so it will be on stand by for future tests of emails.